pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "revvlink"
dynamic = ["version"]
description = "A robust and powerful, fully asynchronous Lavalink wrapper built for discord.py in Python."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name="@JustNixx and @IamGroot", email="nksmay13@gmail.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "aiohttp>=3.9.0,<4",
    "discord.py>=2.4.0",
    "yarl>=1.9.4",
    "typing_extensions>=4.5.0",
    "async_timeout; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "pytest-asyncio",
    "pytest-cov",
    "coverage",
    "aioresponses",
    "pre-commit"
]
redis = ["aioredis>=2"]
hikari = ["hikari>=2.0"]
docs = ["mkdocs-material", "mkdocstrings[python]"]
dave = ["davey>=1.0.0"]

[project.urls]
"Homepage" = "https://github.com/ded-lmfao/RevvLink"

[tool.hatch.version]
path = "revvlink/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["revvlink"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "ASYNC", "PYI", "RUF", "TCH", "PTH"]

[tool.ruff.lint.isort]
known-first-party = ["revvlink"]

[tool.ruff.lint.per-file-ignores]
"revvlink/__init__.py" = ["F403"]
"docs/conf.py" = ["PTH"]
"docs/hooks.py" = ["RUF001"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "strict"
exclude = [
    "docs/**",
    "build/**",
    "tests/**",
    "**/node_modules",
    "**/__pycache__", 
    "**/.*"
]
reportMissingImports = true
reportPrivateUsage = false
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUntypedBaseClass = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
omit = ["tests/*"]
relative_files = true

[tool.coverage.report]
fail_under = 80
